requirements.txt
flask>=3.1.2
flask-sqlalchemy>=3.1.1
flask-socketio>=5.5.1
openpyxl>=3.1.5
pandas>=2.3.3
pillow>=12.0.0
python-socketio>=5.14.2
qrcode>=8.2
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
eventlet>=0.35.2
gunicorn>=21.2.0
b2sdk>=2.5.0
reportlab>=4.0.0